requirements.txt
python-dotenv==0.15.0
django-allauth==0.44.0
gspread==3.6.0
google-auth==1.24.0
Faker~=5.5.0
Pillow==8.1.0
freezegun==1.1.0
